♻️ Extract core from cli into standalone package
