+++ packages/cli/src/constants.ts
@@ -1,0 +1,1 @@
+export const NAME = "lokse";
