--- packages/cli/src/core/reader/index.ts
@@ -1,5 +1,0 @@
-export { default, WorksheetLinesByTitle } from "./spreadsheet-reader";
-export {
-  default as WorksheetReader,
-  InvalidFilterError,
-} from "./worksheet-reader";
