+++ packages/core/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@lokse/core",
+  "description": "Core of localization from spreadsheet solution lokse",
+  "version": "1.5.2",
+  "author": {
+    "name": "Lukas Horak",
+    "email": "horaklukas@centrum.cz"
+  },
+  "bugs": "https://github.com/AckeeCZ/lokse/issues",
+  "dependencies": {
+    "@oclif/errors": "^1.3.4",
+    "@types/bluebird": "^3.5.32",
+    "@types/dedent": "^0.7.0",
+    "@types/google-spreadsheet": "^3.0.1",
+    "@types/mkdirp": "^1.0.1",
+    "bluebird": "^3.7.2",
+    "dedent": "^0.7.0",
+    "google-spreadsheet": "^3.0.13",
+    "lodash": "^4.17.20",
+    "mkdirp": "^1.0.4",
+    "prettier": "^2.1.2"
+  },
+  "devDependencies": {
+    "@types/node": "^10",
+    "typescript": "^3.3"
+  },
+  "engines": {
+    "node": ">=8.0.0"
+  },
+  "homepage": "https://github.com/AckeeCZ/lokse",
+  "keywords": [
+    "lokse"
+  ],
+  "license": "MIT",
+  "main": "lib/index.js",
+  "repository": "AckeeCZ/lokse",
+  "scripts": {
+    "build": "tsc -b"
+  }
+}
