+++ packages/core/src/constants.ts
@@ -1,5 +1,3 @@
-export const NAME = "lokse";
-
 export enum OutputFormat {
   JSON = "json",
   ANDROID = "android",
