+++ packages/core/src/reader/index.ts
@@ -1,0 +1,8 @@
+export { default as Reader } from "./spreadsheet-reader";
+export type { WorksheetLinesByTitle } from "./spreadsheet-reader";
+
+export {
+  default as WorksheetReader,
+  InvalidFilterError,
+} from "./worksheet-reader";
+export type { SheetsFilter } from "./worksheet-reader";
