+++ packages/core/src/utils.ts
@@ -1,5 +1,3 @@
-export * from "./invariants";
-
 export const forceArray = <T>(val: T | T[] | null) => {
   if (Array.isArray(val)) return val;
   if (!val) return [];
